+++ R/is_model_supported.R
@@ -123,6 +123,7 @@ supported_models <- function() {
     "logistf",
     "LORgee",
     "lrm",
+    "manova",
     "MANOVA",
     "maxLik",
     "mcmc",